rename mag_tag to magtag. Remove neopixel power pin handling from magtag example
